# 🚀 Critical: Production-Ready Build for Thursday Jury Launch