--- a/Plugin/assets/css/mt-modal-fix.css
+++ b/Plugin/assets/css/mt-modal-fix.css
@@ -1,56 +1,54 @@
 /**
  * Modal Fix for Mobility Trailblazers
  * Ensures modals are properly centered and visible
+ * Uses consolidated z-index system
  */
 
 /* Override any conflicting styles and ensure proper modal display */
 .mt-modal {
-    position: fixed !important;
-    top: 0 !important;
-    left: 0 !important;
-    right: 0 !important;
-    bottom: 0 !important;
-    width: 100% !important;
-    height: 100% !important;
-    background-color: rgba(0, 0, 0, 0.6) !important;
-    z-index: 999999 !important;
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0, 0, 0, 0.6);
+    z-index: var(--mt-z-modal-backdrop, 400);
     display: none;
-    align-items: center !important;
-    justify-content: center !important;
-    overflow: auto !important;
+    align-items: center;
+    justify-content: center;
+    overflow: auto;
 }
 
 /* When modal is shown - handle different display methods */
 .mt-modal[style*="display: flex"],
 .mt-modal[style*="display:flex"],
 .mt-modal.show {
-    display: flex !important;
-    align-items: center !important;
-    justify-content: center !important;
+    display: flex;
+    align-items: center;
+    justify-content: center;
 }
 
-/* Modal content styling - force center positioning */
+/* Modal content styling - center positioning */
 .mt-modal .mt-modal-content {
-    position: fixed !important;
-    top: 50% !important;
-    left: 50% !important;
-    transform: translate(-50%, -50%) !important;
-    background: #ffffff !important;
-    padding: 30px !important;
-    max-width: 600px !important;
-    width: 90% !important;
-    max-height: 90vh !important;
-    overflow-y: auto !important;
-    border-radius: 8px !important;
-    box-shadow: 0 10px 40px rgba(0,0,0,0.2) !important;
-    z-index: 1000000 !important;
-    margin: 0 !important;
+    position: relative;
+    background: #ffffff;
+    padding: 30px;
+    max-width: 600px;
+    width: 90%;
+    max-height: 90vh;
+    overflow-y: auto;
+    border-radius: 8px;
+    box-shadow: 0 10px 40px rgba(0,0,0,0.2);
+    z-index: var(--mt-z-modal-content, 420);
+    margin: 0;
+    transform: none;
 }
 
-/* Ensure modal content is above everything */
+/* Ensure modal content is properly layered */
 .mt-modal .mt-modal-content * {
     position: relative;
-    z-index: 1000000;
 }
 
 /* Close button styling */
@@ -99,7 +97,7 @@
 
 /* Ensure WordPress admin bar doesn't interfere */
 body.admin-bar .mt-modal {
-    top: 0 !important;
+    top: 0;
 }
 
 /* Fix for scrolled content */