--- a/Plugin/assets/js/candidate-import.js
+++ b/Plugin/assets/js/candidate-import.js
@@ -17,7 +17,7 @@ jQuery(document).ready(function($) {
         e.preventDefault();
         // Double-check mt_ajax exists
         if (typeof mt_ajax === 'undefined') {
-            alert('Import functionality is not properly initialized. Please refresh the page and try again.');
+            alert(mt_ajax.i18n?.import_not_initialized || 'Import functionality is not properly initialized. Please refresh the page and try again.');
             return;
         }
         // Create a temporary file input
@@ -42,9 +42,9 @@ jQuery(document).ready(function($) {
             if (!fileName.endsWith('.csv')) {
                 var fileExt = fileName.split('.').pop();
                 if (fileExt === 'xlsx' || fileExt === 'xls') {
-                    alert('Please convert your Excel file to CSV format first. Use "Save As" in Excel and choose "CSV (Comma delimited)" as the file type.');
+                    alert(mt_ajax.i18n?.excel_to_csv_message || 'Please convert your Excel file to CSV format first. Use "Save As" in Excel and choose "CSV (Comma delimited)" as the file type.');
                 } else {
-                    alert(mt_ajax.i18n.invalid_file_type || 'Please select a CSV file.');
+                    alert(mt_ajax.i18n?.invalid_file_type || 'Please select a CSV file.');
                 }
                 fileInput.remove();
                 return;
@@ -205,9 +205,9 @@ jQuery(document).ready(function($) {
                 if (!fileName.endsWith('.csv')) {
                     var fileExt = fileName.split('.').pop();
                     if (fileExt === 'xlsx' || fileExt === 'xls') {
-                        alert('Please convert your Excel file to CSV format first. Use "Save As" in Excel and choose "CSV (Comma delimited)" as the file type.');
+                        alert(mt_ajax.i18n?.excel_to_csv_message || 'Please convert your Excel file to CSV format first. Use "Save As" in Excel and choose "CSV (Comma delimited)" as the file type.');
                     } else {
-                        alert(mt_ajax.i18n.invalid_file_type || 'Please select a CSV file.');
+                        alert(mt_ajax.i18n?.invalid_file_type || 'Please select a CSV file.');
                     }
                     return;
                 }