… AUDIT_LOG.md
… DATA_INTEGRITY.md
… DECISIONS_LOG.md
… FRONTEND_AUDIT.md
… JS_CLEANUP.md
… NIGHTLY_BUILD_SUMMARY.md
… PERFORMANCE_METRICS.md
… PRODUCTION_READY.md
… Plugin/assets/css/mt-critical-fixes-optimized.min.css
… Plugin/assets/css/mt-responsive-fixes.min.css
… Plugin/assets/css/mt-z-index-system.min.css
… Plugin/includes/ajax/class-mt-assignment-ajax.php
… Plugin/includes/ajax/class-mt-csv-import-ajax.php
… Plugin/includes/ajax/class-mt-evaluation-ajax.php
… Plugin/includes/core/class-mt-plugin.php
… Plugin/includes/db/class-mt-database-optimizer.php
… Plugin/includes/repositories/class-mt-evaluation-repository.php
… Plugin/includes/services/class-mt-evaluation-service.php
… Plugin/includes/validators/class-mt-data-validator.php
… Plugin/languages/mobility-trailblazers-de_DE.po
… Plugin/templates/admin/assignments.php
… Plugin/templates/admin/evaluations-inline-fix.php
… Plugin/templates/frontend/candidates-grid.php
… Plugin/templates/frontend/single/single-mt_jury.php
… SECURITY_REPORT.md
… TRANSLATION_REPORT.md
… USER_FLOW_RESULTS.md
… compile-mo.php
… docs/css-audit/stylelint-report.json
… docs/css-audit/stylelint-verbose.txt